+++ newton/_src/sensors/warp_raytrace/bvh.py
@@ -149,7 +149,6 @@ def compute_shape_bvh_bounds(
     shape_world_index: wp.array(dtype=wp.int32),
     shape_enabled: wp.array(dtype=wp.uint32),
     shape_types: wp.array(dtype=wp.int32),
-    shape_indices: wp.array(dtype=wp.int32),
     shape_sizes: wp.array(dtype=wp.vec3f),
     shape_transforms: wp.array(dtype=wp.transformf),
     shape_bounds: wp.array2d(dtype=wp.vec3f),
@@ -193,8 +192,8 @@ def compute_shape_bvh_bounds(
     elif geom_type == GeoType.BOX:
         lower, upper = compute_box_bounds(transform, size)
     elif geom_type == GeoType.MESH or geom_type == GeoType.GAUSSIAN:
-        min_bounds = shape_bounds[shape_indices[shape_index], 0]
-        max_bounds = shape_bounds[shape_indices[shape_index], 1]
+        min_bounds = shape_bounds[shape_index, 0]
+        max_bounds = shape_bounds[shape_index, 1]
         lower, upper = compute_shape_bounds(transform, size, min_bounds, max_bounds)
 
     out_bvh_lowers[bvh_index_local] = lower