+++ newton/_src/sensors/warp_raytrace/raytrace.py
@@ -21,6 +21,7 @@ import warp as wp
 
 from ...geometry import Gaussian, GeoType, raycast
 from . import gaussians, ray_intersect
+from .types import MeshData
 
 if TYPE_CHECKING:
     from .render_context import RenderContext
@@ -40,7 +41,6 @@ class ClosestHit:
     bary_u: wp.float32
     bary_v: wp.float32
     face_idx: wp.int32
-    shape_mesh_index: wp.int32
     color: wp.vec3f
 
 
@@ -65,10 +65,11 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
         world_index: wp.int32,
         shape_enabled: wp.array(dtype=wp.uint32),
         shape_types: wp.array(dtype=wp.int32),
-        shape_indices: wp.array(dtype=wp.int32),
         shape_sizes: wp.array(dtype=wp.vec3f),
         shape_transforms: wp.array(dtype=wp.transformf),
         shape_source_ptr: wp.array(dtype=wp.uint64),
+        shape_mesh_data_ids: wp.array(dtype=wp.int32),
+        mesh_data: wp.array(dtype=MeshData),
         gaussians_data: wp.array(dtype=Gaussian.Data),
         ray_origin_world: wp.vec3f,
         ray_dir_world: wp.vec3f,
@@ -93,17 +94,17 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
                     hit_u = wp.float32(0.0)
                     hit_v = wp.float32(0.0)
                     hit_face_id = wp.int32(-1)
-                    hit_mesh_id = wp.int32(-1)
                     hit_color = wp.vec3f(0.0)
 
                     if shape_types[si] == GeoType.MESH:
-                        hit_mesh_id = shape_indices[si]
                         geom_hit, hit_u, hit_v, hit_face_id = ray_intersect.ray_intersect_mesh(
                             shape_transforms[si],
                             shape_sizes[si],
                             ray_origin_world,
                             ray_dir_world,
-                            shape_source_ptr[hit_mesh_id],
+                            shape_source_ptr[si],
+                            shape_mesh_data_ids[si],
+                            mesh_data,
                             wp.static(config.enable_backface_culling),
                             closest_hit.distance,
                         )
@@ -161,7 +162,7 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
                         if num_gaussians_hit < wp.static(state.num_gaussians):
                             gaussians_hit[num_gaussians_hit] = si
                             num_gaussians_hit += 1
-                            # gaussian_id = shape_source_ptr[shape_indices[si]]
+                            # gaussian_id = shape_source_ptr[si]
                             # geom_hit, hit_color = shade_gaussians(
                             #     shape_transforms[si],
                             #     shape_sizes[si],
@@ -178,7 +179,6 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
                         closest_hit.bary_u = hit_u
                         closest_hit.bary_v = hit_v
                         closest_hit.face_idx = hit_face_id
-                        closest_hit.shape_mesh_index = hit_mesh_id
                         closest_hit.color = hit_color
 
                 # Temporary workaround. Warp BVH queries share some stack data,
@@ -192,7 +192,7 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
                     for gi in range(num_gaussians_hit):
                         si = gaussians_hit[gi]
 
-                        gaussian_id = shape_source_ptr[shape_indices[si]]
+                        gaussian_id = shape_source_ptr[si]
                         geom_hit, hit_color = shade_gaussians(
                             shape_transforms[si],
                             shape_sizes[si],
@@ -209,7 +209,6 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
                             closest_hit.bary_u = hit_u
                             closest_hit.bary_v = hit_v
                             closest_hit.face_idx = hit_face_id
-                            closest_hit.shape_mesh_index = hit_mesh_id
                             closest_hit.color = hit_color
 
         return closest_hit
@@ -247,7 +246,6 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
                         closest_hit.distance = geom_hit.distance
                         closest_hit.normal = geom_hit.normal
                         closest_hit.shape_index = PARTICLES_SHAPE_ID
-                        closest_hit.shape_mesh_index = -1
 
         return closest_hit
 
@@ -273,7 +271,6 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
                 closest_hit.bary_u = bary_u
                 closest_hit.bary_v = bary_v
                 closest_hit.face_idx = face_idx
-                closest_hit.shape_mesh_index = -1
 
         return closest_hit
 
@@ -289,10 +286,11 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
         max_distance: wp.float32,
         shape_enabled: wp.array(dtype=wp.uint32),
         shape_types: wp.array(dtype=wp.int32),
-        shape_indices: wp.array(dtype=wp.int32),
         shape_sizes: wp.array(dtype=wp.vec3f),
         shape_transforms: wp.array(dtype=wp.transformf),
         shape_source_ptr: wp.array(dtype=wp.uint64),
+        shape_mesh_data_ids: wp.array(dtype=wp.int32),
+        mesh_data: wp.array(dtype=MeshData),
         particles_position: wp.array(dtype=wp.vec3f),
         particles_radius: wp.array(dtype=wp.float32),
         triangle_mesh_id: wp.uint64,
@@ -307,7 +305,6 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
         closest_hit.bary_u = wp.float32(0.0)
         closest_hit.bary_v = wp.float32(0.0)
         closest_hit.face_idx = wp.int32(-1)
-        closest_hit.shape_mesh_index = wp.int32(-1)
         closest_hit.color = wp.vec3f(0.0)
 
         closest_hit = closest_hit_triangle_mesh(closest_hit, triangle_mesh_id, ray_origin_world, ray_dir_world)
@@ -320,10 +317,11 @@ def create_closest_hit_function(config: RenderContext.Config, state: RenderConte
             world_index,
             shape_enabled,
             shape_types,
-            shape_indices,
             shape_sizes,
             shape_transforms,
             shape_source_ptr,
+            shape_mesh_data_ids,
+            mesh_data,
             gaussians_data,
             ray_origin_world,
             ray_dir_world,
@@ -356,7 +354,6 @@ def create_first_hit_function(config: RenderContext.Config, state: RenderContext
         world_index: wp.int32,
         shape_enabled: wp.array(dtype=wp.uint32),
         shape_types: wp.array(dtype=wp.int32),
-        shape_indices: wp.array(dtype=wp.int32),
         shape_sizes: wp.array(dtype=wp.vec3f),
         shape_transforms: wp.array(dtype=wp.transformf),
         shape_source_ptr: wp.array(dtype=wp.uint64),
@@ -379,17 +376,14 @@ def create_first_hit_function(config: RenderContext.Config, state: RenderContext
                     dist = wp.float32(-1)
 
                     if shape_types[si] == GeoType.MESH:
-                        geom_hit, _u, _v, _f = ray_intersect.ray_intersect_mesh(
+                        dist = ray_intersect.ray_intersect_mesh_any(
                             shape_transforms[si],
                             shape_sizes[si],
                             ray_origin_world,
                             ray_dir_world,
-                            shape_source_ptr[shape_indices[si]],
-                            wp.static(config.enable_backface_culling),
+                            shape_source_ptr[si],
                             max_dist,
                         )
-                        if geom_hit.hit:
-                            dist = geom_hit.distance
                     elif shape_types[si] == GeoType.PLANE:
                         dist = ray_intersect.ray_intersect_plane(
                             shape_transforms[si],
@@ -498,7 +492,6 @@ def create_first_hit_function(config: RenderContext.Config, state: RenderContext
         world_index: wp.int32,
         shape_enabled: wp.array(dtype=wp.uint32),
         shape_types: wp.array(dtype=wp.int32),
-        shape_indices: wp.array(dtype=wp.int32),
         shape_sizes: wp.array(dtype=wp.vec3f),
         shape_transforms: wp.array(dtype=wp.transformf),
         shape_source_ptr: wp.array(dtype=wp.uint64),
@@ -519,7 +512,6 @@ def create_first_hit_function(config: RenderContext.Config, state: RenderContext
             world_index,
             shape_enabled,
             shape_types,
-            shape_indices,
             shape_sizes,
             shape_transforms,
             shape_source_ptr,